pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastmcp>=2.11.2",
    "httpx>=0.28.1",
    "langchain-groq>=0.3.7",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.4",
    "markdownify>=1.1.0",
    "mcp>=1.12.4",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
]
